Successfully created first CCE node
Upgrade the "opentelekomcloud" provider to 1.30.2.

Use simpler CIDR block.

Remove superfluous CCE cluster default parameter.

Add a SSH public key which we can use when creating nodes. (I tried
to use a key pair which I had generated in the web console, but Terraform could
not see it. Conversely, this key pair is not visible in the web console,
but it works.)